requirements.txt
python-telegram-bot==21.6
APScheduler==3.10.4
aiohttp==3.11.11
aiosqlite==0.20.0
python-dotenv==1.0.1
